+++ src/Seomatic.php
@@ -1,12 +1,12 @@
 <?php
 /**
- * SEOmatic plugin for Craft CMS 3.x
+ * SEOmatic plugin for Craft CMS
  *
  * A turnkey SEO implementation for Craft CMS that is comprehensive, powerful,
  * and flexible
  *
  * @link      https://nystudio107.com
- * @copyright Copyright (c) 2017 nystudio107
+ * @copyright Copyright (c) 2022 nystudio107
  */
 
 namespace nystudio107\seomatic;
@@ -48,8 +48,6 @@ use markhuot\CraftQL\Builders\Schema;
 use markhuot\CraftQL\CraftQL;
 use markhuot\CraftQL\Events\AlterSchemaFields;
 use nystudio107\fastcgicachebust\FastcgiCacheBust;
-use nystudio107\pluginmanifest\services\ManifestService;
-use nystudio107\seomatic\assetbundles\seomatic\SeomaticAsset;
 use nystudio107\seomatic\autocompletes\TrackingVarsAutocomplete;
 use nystudio107\seomatic\fields\Seomatic_Meta as Seomatic_MetaField;
 use nystudio107\seomatic\fields\SeoSettings as SeoSettingsField;
@@ -66,17 +64,7 @@ use nystudio107\seomatic\integrations\feedme\SeoSettings as SeoSettingsFeedMe;
 use nystudio107\seomatic\listeners\GetCraftQLSchema;
 use nystudio107\seomatic\models\MetaScriptContainer;
 use nystudio107\seomatic\models\Settings;
-use nystudio107\seomatic\services\FrontendTemplates as FrontendTemplatesService;
-use nystudio107\seomatic\services\Helper as HelperService;
-use nystudio107\seomatic\services\JsonLd as JsonLdService;
-use nystudio107\seomatic\services\Link as LinkService;
-use nystudio107\seomatic\services\MetaBundles as MetaBundlesService;
-use nystudio107\seomatic\services\MetaContainers as MetaContainersService;
-use nystudio107\seomatic\services\Script as ScriptService;
-use nystudio107\seomatic\services\SeoElements as SeoElementsService;
-use nystudio107\seomatic\services\Sitemaps as SitemapsService;
-use nystudio107\seomatic\services\Tag as TagService;
-use nystudio107\seomatic\services\Title as TitleService;
+use nystudio107\seomatic\services\ServicesTrait;
 use nystudio107\seomatic\twigextensions\SeomaticTwigExtension;
 use nystudio107\seomatic\variables\SeomaticVariable;
 use nystudio107\twigfield\autocompletes\EnvironmentVariableAutocomplete;
@@ -94,22 +82,14 @@ use yii\base\Event;
  * @author    nystudio107
  * @package   Seomatic
  * @since     3.0.0
- *
- * @property FrontendTemplatesService $frontendTemplates
- * @property HelperService $helper
- * @property JsonLdService $jsonLd
- * @property LinkService $link
- * @property MetaBundlesService $metaBundles
- * @property MetaContainersService $metaContainers
- * @property ScriptService $script
- * @property SeoElementsService $seoElements
- * @property SitemapsService $sitemaps
- * @property TagService $tag
- * @property TitleService $title
- * @property ManifestService $manifest
  */
 class Seomatic extends Plugin
 {
+    // Traits
+    // =========================================================================
+
+    use ServicesTrait;
+
     // Constants
     // =========================================================================
 
@@ -230,8 +210,6 @@ class Seomatic extends Plugin
      */
     public static $craft37 = false;
 
-    // Static Methods
-    // =========================================================================
     /**
      * @var string
      */
@@ -243,39 +221,14 @@ class Seomatic extends Plugin
 
     // Public Properties
     // =========================================================================
+
     /**
      * @var bool
      */
     public $hasCpSettings = true;
 
-    /**
-     * @inheritdoc
-     */
-    public function __construct($id, $parent = null, array $config = [])
-    {
-        $config['components'] = [
-            'frontendTemplates' => FrontendTemplatesService::class,
-            'helper' => HelperService::class,
-            'jsonLd' => JsonLdService::class,
-            'link' => LinkService::class,
-            'metaBundles' => MetaBundlesService::class,
-            'metaContainers' => MetaContainersService::class,
-            'script' => ScriptService::class,
-            'seoElements' => SeoElementsService::class,
-            'sitemaps' => SitemapsService::class,
-            'tag' => TagService::class,
-            'title' => TitleService::class,
-            // Register the manifest service
-            'manifest' => [
-                'class' => ManifestService::class,
-                'assetClass' => SeomaticAsset::class,
-                'devServerManifestPath' => 'http://craft-seomatic-buildchain:8080/',
-                'devServerPublicPath' => 'http://craft-seomatic-buildchain:8080/',
-            ],
-        ];
-
-        parent::__construct($id, $parent, $config);
-    }
+    // Static Methods
+    // =========================================================================
 
     /**
      * Set the matched element